refactor: move fmt_local_time to cache/utils.py
Formatting logic has no dependency on the DB and was misplaced in the
storage module. Extracted alongside the shared LOCAL_TZ setup into a
new cache/utils.py; store.py and server.py import from there instead.
